+++ kbl/kbl-v25/src/main/resources/kbl25/KBL25_SR1-ext.xjb
@@ -86,7 +86,7 @@
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Processing_information']]">
-        <inheritance:implements>com.foursoft.harness.kbl.common.HasProcessingInformation&lt;com.foursoft.harness.kbl.v25.KblProcessingInstruction&gt;</inheritance:implements>
+        <inheritance:implements>com.foursoft.harness.kbl.v25.HasProcessingInformation</inheritance:implements>
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Reference_element']]">
